+++ QuanLyKhachSan/Form/frm_DichVu.cs
@@ -162,13 +162,11 @@ namespace QuanLyKhachSan
                 string maPhong = cbo_MaPhong.SelectedItem.ToString();
                 string maDichVu = cbo_MaDichVu3.SelectedValue.ToString();
 
-                // Kiểm tra xem dịch vụ đã tồn tại trong ListView chưa
                 bool daTonTai = false;
                 foreach (ListViewItem existingItem in listView_DichVuDaDat.Items)
                 {
                     if (existingItem.SubItems[0].Text == maDichVu && existingItem.SubItems[2].Text == maPhong)
                     {
-                        // Cập nhật số lượng của dịch vụ đã tồn tại
                         existingItem.SubItems[1].Text = (int.Parse(existingItem.SubItems[1].Text) + soLuong).ToString();
                         daTonTai = true;
                         break;
@@ -177,7 +175,6 @@ namespace QuanLyKhachSan
 
                 if (!daTonTai)
                 {
-                    // Thêm thông tin vào ListView
                     ListViewItem item = new ListViewItem(new string[] { maDichVu, soLuong.ToString(), maPhong });
                     listView_DichVuDaDat.Items.Add(item);
                 }
@@ -219,7 +216,6 @@ namespace QuanLyKhachSan
         }
         private string GetTenDichVuByMaDichVu(string tenDichVu)
         {
-            // Gọi phương thức của bus để lấy tên dịch vụ từ mã dịch vụ
             return bus.GetTenDichVuByMaDichVu(tenDichVu);
         }
 
@@ -234,41 +230,36 @@ namespace QuanLyKhachSan
                 }
                 foreach (ListViewItem item in listView_DichVuDaDat.Items)
                 {
-                    string tenDichVu = item.SubItems[0].Text; // Tên dịch vụ được lưu trong cột đầu tiên
-                    string maPhong = item.SubItems[2].Text;   // Mã phòng được lưu trong cột thứ ba
+                    string tenDichVu = item.SubItems[0].Text; 
+                    string maPhong = item.SubItems[2].Text;  
                     int soLuong = int.Parse(item.SubItems[1].Text);
 
-                    string maDichVu = bus.GetMaDichVuByTenDichVu(tenDichVu); // Lấy mã dịch vụ từ tên dịch vụ
-                    float donGia = bus.GetDonGiaByMaDichVu(maDichVu); // Lấy đơn giá từ mã dịch vụ
+                    string maDichVu = bus.GetMaDichVuByTenDichVu(tenDichVu); 
+                    double donGia = bus.GetDonGiaByMaDichVu(maDichVu); 
 
-                    // Kiểm tra xem dịch vụ đã tồn tại trong cùng một phòng chưa
+                  
                     if (bus.CheckDichVuDaTonTai(maDichVu, maPhong))
                     {
-                        // Nếu đã tồn tại, cập nhật số lượng dịch vụ đã đăng ký
+                     
                         bus.UpdateDangKyDichVu(maDichVu, maPhong, soLuong);
                     }
                     else
                     {
-                        // Nếu chưa tồn tại, thêm mới dịch vụ vào cơ sở dữ liệu
-                        // Tạo mã đăng ký dịch vụ ngẫu nhiên
+                       
                         Random random = new Random();
                         string maDangKyDV = "DDV" + random.Next(100, 999).ToString();
 
-                        // Kiểm tra xem mã đăng ký dịch vụ đã tồn tại chưa
                         while (bus.CheckMaDangKyDVTonTai(maDangKyDV))
                         {
                             maDangKyDV = "DDV" + random.Next(100, 999).ToString();
                         }
 
-                        // Thêm thông tin vào bảng DangKyDichVu
                         bus.InsertDangKyDichVu(maDangKyDV, maDichVu, maPhong, soLuong, donGia);
                     }
 
-                    // Gọi phương thức của form frm_ThongTinPhong để cập nhật thông tin dịch vụ đã đăng ký
                     frm_ThongTinPhong.Instance.UpdateDichVuDaSuDung(item.Clone() as ListViewItem); // Tạo bản sao của item và thêm vào ListView khác
                 }
                 MessageBox.Show("Đăng ký dịch vụ thành công!", "Thông Báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                // Clear ListView sau khi đăng ký thành công
                 listView_DichVuDaDat.Items.Clear();
             }
             catch (Exception ex)
@@ -278,7 +269,6 @@ namespace QuanLyKhachSan
         }
         private string GetMaDichVuByTenDichVu(string tenDichVu)
         {
-            // Gọi phương thức của bus để lấy mã dịch vụ từ tên dịch vụ
             return bus.GetMaDichVuByTenDichVu(tenDichVu);
         }
 
@@ -295,7 +285,6 @@ namespace QuanLyKhachSan
             {
                 string maDichVu = dgv_DanhSachDichVu3.Rows[e.RowIndex].Cells["MaDichVu"].Value.ToString();
 
-                // Lặp qua danh sách tùy chọn trong combobox để tìm giá trị khớp
                 for (int i = 0; i < cbo_MaDichVu3.Items.Count; i++)
                 {
                     if (cbo_MaDichVu3.GetItemText(cbo_MaDichVu3.Items[i]) == maDichVu)
